Move dump code into datadumper
